+++ readthedocs/core/templatetags/privacy_tags.py
@@ -1,11 +1,12 @@
 """Template tags to query projects by privacy."""
 
 from __future__ import absolute_import
+
 from django import template
 
-from ..loader import AdminPermission
-
+from readthedocs.core.permissions import AdminPermission
 from readthedocs.projects.models import Project
+
 
 register = template.Library()
