--- readthedocs/oauth/managers.py
@@ -1,12 +1,0 @@
-"""Managers for OAuth models"""
-
-from __future__ import absolute_import
-from readthedocs.privacy.loader import RelatedUserQuerySet
-
-
-class RemoteRepositoryQuerySet(RelatedUserQuerySet):
-    pass
-
-
-class RemoteOrganizationQuerySet(RelatedUserQuerySet):
-    pass
