--- readthedocs/privacy/backend.py
@@ -1,326 +1,0 @@
-"""Django Managers and Querysets to apply project privacy restrictions."""
-from __future__ import absolute_import
-from builtins import object
-from django.db import models
-
-from guardian.shortcuts import get_objects_for_user
-
-from readthedocs.builds.constants import BRANCH
-from readthedocs.builds.constants import TAG
-from readthedocs.builds.constants import LATEST
-from readthedocs.builds.constants import LATEST_VERBOSE_NAME
-from readthedocs.builds.constants import STABLE
-from readthedocs.builds.constants import STABLE_VERBOSE_NAME
-from readthedocs.core.utils.extend import (SettingsOverrideObject,
-                                           get_override_class)
-from readthedocs.projects import constants
-
-
-class ProjectQuerySet(models.QuerySet):
-
-    """Projects take into account their own privacy_level setting."""
-
-    use_for_related_fields = True
-
-    def _add_user_repos(self, queryset, user):
-        if user.has_perm('projects.view_project'):
-            return self.all().distinct()
-        if user.is_authenticated():
-            user_queryset = get_objects_for_user(user, 'projects.view_project')
-            queryset = user_queryset | queryset
-        return queryset.distinct()
-
-    def for_user_and_viewer(self, user, viewer):
-        """Show projects that a user owns, that another user can see."""
-        queryset = self.filter(privacy_level=constants.PUBLIC)
-        queryset = self._add_user_repos(queryset, viewer)
-        queryset = queryset.filter(users__in=[user])
-        return queryset
-
-    def for_admin_user(self, user=None):
-        if user.is_authenticated():
-            return self.filter(users__in=[user])
-        return self.none()
-
-    def public(self, user=None):
-        queryset = self.filter(privacy_level=constants.PUBLIC)
-        if user:
-            return self._add_user_repos(queryset, user)
-        return queryset
-
-    def protected(self, user=None):
-        queryset = self.filter(privacy_level__in=[constants.PUBLIC, constants.PROTECTED])
-        if user:
-            return self._add_user_repos(queryset, user)
-        return queryset
-
-    def private(self, user=None):
-        queryset = self.filter(privacy_level=constants.PRIVATE)
-        if user:
-            return self._add_user_repos(queryset, user)
-        return queryset
-
-    # Aliases
-
-    def dashboard(self, user=None):
-        return self.for_admin_user(user)
-
-    def api(self, user=None):
-        return self.public(user)
-
-
-class VersionManager(models.Manager):
-
-    """Version manager for manager only queries
-
-    For queries not suitable for the :py:cls:`VersionQuerySet`, such as create
-    queries.
-    """
-
-    @classmethod
-    def from_queryset(cls, queryset_class, class_name=None):
-        # This is overridden because :py:meth:`models.Manager.from_queryset`
-        # uses `inspect` to retrieve the class methods, and the proxy class has
-        # no direct members.
-        queryset_class = get_override_class(
-            VersionQuerySet,
-            VersionQuerySet._default_class  # pylint: disable=protected-access
-        )
-        return super(VersionManager, cls).from_queryset(queryset_class, class_name)
-
-    def create_stable(self, **kwargs):
-        defaults = {
-            'slug': STABLE,
-            'verbose_name': STABLE_VERBOSE_NAME,
-            'machine': True,
-            'active': True,
-            'identifier': STABLE,
-            'type': TAG,
-        }
-        defaults.update(kwargs)
-        return self.create(**defaults)
-
-    def create_latest(self, **kwargs):
-        defaults = {
-            'slug': LATEST,
-            'verbose_name': LATEST_VERBOSE_NAME,
-            'machine': True,
-            'active': True,
-            'identifier': LATEST,
-            'type': BRANCH,
-        }
-        defaults.update(kwargs)
-        return self.create(**defaults)
-
-
-class VersionQuerySetBase(models.QuerySet):
-
-    """Versions take into account their own privacy_level setting."""
-
-    use_for_related_fields = True
-
-    def _add_user_repos(self, queryset, user):
-        if user.has_perm('builds.view_version'):
-            return self.all().distinct()
-        if user.is_authenticated():
-            user_queryset = get_objects_for_user(user, 'builds.view_version')
-            queryset = user_queryset | queryset
-        return queryset.distinct()
-
-    def public(self, user=None, project=None, only_active=True):
-        queryset = self.filter(privacy_level=constants.PUBLIC)
-        if user:
-            queryset = self._add_user_repos(queryset, user)
-        if project:
-            queryset = queryset.filter(project=project)
-        if only_active:
-            queryset = queryset.filter(active=True)
-        return queryset
-
-    def protected(self, user=None, project=None, only_active=True):
-        queryset = self.filter(privacy_level__in=[constants.PUBLIC, constants.PROTECTED])
-        if user:
-            queryset = self._add_user_repos(queryset, user)
-        if project:
-            queryset = queryset.filter(project=project)
-        if only_active:
-            queryset = queryset.filter(active=True)
-        return queryset
-
-    def private(self, user=None, project=None, only_active=True):
-        queryset = self.filter(privacy_level__in=[constants.PRIVATE])
-        if user:
-            queryset = self._add_user_repos(queryset, user)
-        if project:
-            queryset = queryset.filter(project=project)
-        if only_active:
-            queryset = queryset.filter(active=True)
-        return queryset
-
-    def api(self, user=None):
-        return self.public(user, only_active=False)
-
-    def for_project(self, project):
-        """Return all versions for a project, including translations"""
-        return self.filter(
-            models.Q(project=project) |
-            models.Q(project__main_language_project=project)
-        )
-
-
-class VersionQuerySet(SettingsOverrideObject):
-
-    _default_class = VersionQuerySetBase
-
-
-class BuildQuerySet(models.QuerySet):
-
-    """
-    Build objects that are privacy aware.
-
-    i.e. they take into account the privacy of the Version that they relate to.
-    """
-
-    use_for_related_fields = True
-
-    def _add_user_repos(self, queryset, user=None):
-        if user.has_perm('builds.view_version'):
-            return self.all().distinct()
-        if user.is_authenticated():
-            user_queryset = get_objects_for_user(user, 'builds.view_version')
-            pks = [p.pk for p in user_queryset]
-            queryset = self.filter(version__pk__in=pks) | queryset
-        return queryset.distinct()
-
-    def public(self, user=None, project=None):
-        queryset = self.filter(version__privacy_level=constants.PUBLIC)
-        if user:
-            queryset = self._add_user_repos(queryset, user)
-        if project:
-            queryset = queryset.filter(project=project)
-        return queryset
-
-    def api(self, user=None):
-        return self.public(user)
-
-
-class RelatedProjectQuerySet(models.QuerySet):
-
-    """
-    A manager for things that relate to Project and need to get their perms from the project.
-
-    This shouldn't be used as a subclass.
-    """
-
-    use_for_related_fields = True
-    project_field = 'project'
-
-    def _add_user_repos(self, queryset, user=None):
-        # Hack around get_objects_for_user not supporting global perms
-        if user.has_perm('projects.view_project'):
-            return self.all().distinct()
-        if user.is_authenticated():
-            # Add in possible user-specific views
-            project_qs = get_objects_for_user(user, 'projects.view_project')
-            pks = [p.pk for p in project_qs]
-            kwargs = {'%s__pk__in' % self.project_field: pks}
-            queryset = self.filter(**kwargs) | queryset
-        return queryset.distinct()
-
-    def public(self, user=None, project=None):
-        kwargs = {'%s__privacy_level' % self.project_field: constants.PUBLIC}
-        queryset = self.filter(**kwargs)
-        if user:
-            queryset = self._add_user_repos(queryset, user)
-        if project:
-            queryset = queryset.filter(project=project)
-        return queryset
-
-    def protected(self, user=None, project=None):
-        kwargs = {
-            '%s__privacy_level__in' % self.project_field: [constants.PUBLIC, constants.PROTECTED]
-        }
-        queryset = self.filter(**kwargs)
-        if user:
-            queryset = self._add_user_repos(queryset, user)
-        if project:
-            queryset = queryset.filter(project=project)
-        return queryset
-
-    def private(self, user=None, project=None):
-        kwargs = {
-            '%s__privacy_level' % self.project_field: constants.PRIVATE,
-        }
-        queryset = self.filter(**kwargs)
-        if user:
-            queryset = self._add_user_repos(queryset, user)
-        if project:
-            queryset = queryset.filter(project=project)
-        return queryset
-
-    def api(self, user=None):
-        return self.public(user)
-
-
-class ParentRelatedProjectQuerySet(RelatedProjectQuerySet):
-    project_field = 'parent'
-    use_for_related_fields = True
-
-
-class ChildRelatedProjectQuerySet(RelatedProjectQuerySet):
-    project_field = 'child'
-    use_for_related_fields = True
-
-
-class RelatedBuildQuerySet(models.QuerySet):
-
-    """For models with association to a project through :py:class:`Build`"""
-
-    use_for_related_fields = True
-
-    def _add_user_repos(self, queryset, user=None):
-        if user.has_perm('builds.view_version'):
-            return self.all().distinct()
-        if user.is_authenticated():
-            user_queryset = get_objects_for_user(user, 'builds.view_version')
-            pks = [p.pk for p in user_queryset]
-            queryset = self.filter(
-                build__version__pk__in=pks) | queryset
-        return queryset.distinct()
-
-    def public(self, user=None, project=None):
-        queryset = self.filter(build__version__privacy_level=constants.PUBLIC)
-        if user:
-            queryset = self._add_user_repos(queryset, user)
-        if project:
-            queryset = queryset.filter(build__project=project)
-        return queryset
-
-    def api(self, user=None):
-        return self.public(user)
-
-
-class RelatedUserQuerySet(models.QuerySet):
-
-    """For models with relations through :py:class:`User`"""
-
-    def api(self, user=None):
-        """Return objects for user"""
-        if not user.is_authenticated():
-            return self.none()
-        return self.filter(users=user)
-
-
-class AdminPermission(object):
-
-    @classmethod
-    def is_admin(cls, user, project):
-        return user in project.users.all()
-
-    @classmethod
-    def is_member(cls, user, obj):
-        return user in obj.users.all()
-
-
-class AdminNotAuthorized(ValueError):
-    pass
