+++ readthedocs/restapi/permissions.py
@@ -3,7 +3,8 @@
 from __future__ import absolute_import
 
 from rest_framework import permissions
-from readthedocs.privacy.backend import AdminPermission
+
+from readthedocs.core.permissions import AdminPermission
 
 
 class IsOwner(permissions.BasePermission):
